+++ .product/changelog.md
@@ -1,5 +1,8 @@
+## 1.0.3 - Mar 31, 2025
+- Updating .mo files with updated brand name. 
+
 ## 1.0.2 - Mar 31, 2025
-- Immediate soft rebrand to Advanced Category Manager 😆😅
+- Immediate soft rebrand to Advanced Category Manager 
 
 ## 1.0.1 - Mar 31, 2025
 - Initial Release